requirements.txt
flask==3.0.2
google-generativeai==0.3.2
python-dotenv==1.0.1
requests==2.31.0
Pillow==10.2.0
flask-sqlalchemy==3.1.1
flask-login==0.6.3
werkzeug==3.0.1 